A Python game just for training
A simple game whose goal is to challenge players who have the ability to manage their business during a certain period by purchasing real estate and calculating the amount of potential income and taxes during a certain period. The winner is the one with a business mind that balances income, sustainability and smoothness. It seems simple and easy from the outside, but it is not that simple. It requires a lot of planning, strategy and thinking. 

Enjoy